+++ packages/template-ts/src/render/expressions.ts
@@ -42,6 +42,9 @@ export class Evaluator {
       }
       case 'Member':
         return this.runtime.member(this.evaluate(expr.object, frame), expr.key);
+      case 'MemberCall':
+      case 'ClassCall':
+        throw this.runtime.error(frame, expr.span, 'E_RUNTIME_UNKNOWN_FUNCTION', 'object and class function calls are not implemented');
       case 'Index':
         return this.runtime.index(this.evaluate(expr.object, frame), this.evaluate(expr.index, frame));
       case 'Call':